Clear functionality added for the speech to text clipboard
