+++ src/Pages/SpeechToText/TextEditor/Tiptap.jsx
@@ -1,6 +1,7 @@
 import { useEditor, EditorContent } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import Underline from "@tiptap/extension-underline";
+import './TipTap.css'
 import {
     FaBold,
     FaHeading,
